+++ src/main/java/fll/web/admin/GatherJudgeInformation.java
@@ -19,13 +19,15 @@ import java.util.Set;
 import javax.sql.DataSource;
 
 import fll.JudgeInformation;
-import fll.db.Queries;
+import fll.Tournament;
 import fll.scheduler.TournamentSchedule;
 import fll.web.ApplicationAttributes;
 import fll.web.AuthenticationContext;
 import fll.web.BaseFLLServlet;
 import fll.web.SessionAttributes;
+import fll.web.TournamentData;
 import fll.web.UserRole;
+import fll.web.report.awards.AwardsScriptReport;
 import fll.xml.ChallengeDescription;
 import fll.xml.SubjectiveScoreCategory;
 import jakarta.servlet.ServletContext;
@@ -34,7 +36,6 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import net.mtu.eggplant.util.sql.SQLFunctions;
 
 /**
  * Get the information needed to edit judges and store it in the session.
@@ -83,25 +84,24 @@ public class GatherJudgeInformation extends BaseFLLServlet {
 
     final StringBuilder message = new StringBuilder();
     final DataSource datasource = ApplicationAttributes.getDataSource(application);
-    Connection connection = null;
-    try {
-      connection = datasource.getConnection();
-      final int tournament = Queries.getCurrentTournament(connection);
+    final TournamentData tournamentData = ApplicationAttributes.getTournamentData(application);
+    final Tournament tournament = tournamentData.getCurrentTournament();
+    try (Connection connection = datasource.getConnection()) {
       final ChallengeDescription challengeDescription = ApplicationAttributes.getChallengeDescription(application);
 
       final List<SubjectiveScoreCategory> subjectiveCategories = challengeDescription.getSubjectiveCategories();
 
-      if (!TournamentSchedule.scheduleExistsInDatabase(connection, tournament)) {
+      if (!TournamentSchedule.scheduleExistsInDatabase(connection, tournament.getTournamentID())) {
         message.append("<p class='warning'>You have not loaded a schedule. If you intend to do so you should go back to the <a href='index.jsp'>admin page</a> and do this before assigning judges.</p>");
       }
 
-      if (checkForEnteredSubjectiveScores(connection, tournament)) {
+      if (checkForEnteredSubjectiveScores(connection, tournament.getTournamentID())) {
         message.append("<p class='error'>Subjective scores have already been entered for this tournament, changing the judges may cause some scores to be deleted</p>");
       }
 
-      session.setAttribute(JUDGES_KEY, JudgeInformation.getJudges(connection, tournament));
+      session.setAttribute(JUDGES_KEY, JudgeInformation.getJudges(connection, tournament.getTournamentID()));
 
-      session.setAttribute(STATIONS_KEY, gatherJudgingStations(connection, tournament));
+      session.setAttribute(STATIONS_KEY, AwardsScriptReport.getJudgingStationOrder(connection, tournament));
 
       session.setAttribute(CATEGORIES_KEY, gatherCategories(subjectiveCategories));
 
@@ -112,8 +112,6 @@ public class GatherJudgeInformation extends BaseFLLServlet {
     } catch (final SQLException e) {
       LOGGER.error("There was an error talking to the database", e);
       throw new RuntimeException("There was an error talking to the database", e);
-    } finally {
-      SQLFunctions.close(connection);
     }
   }
 
@@ -132,14 +130,6 @@ public class GatherJudgeInformation extends BaseFLLServlet {
         }
       }
     }
-  }
-
-  private List<String> gatherJudgingStations(final Connection connection,
-                                             final int tournament)
-      throws SQLException {
-    final List<String> stations = Queries.getJudgingStations(connection, tournament);
-
-    return stations;
   }
 
   private Map<String, String> gatherCategories(final List<SubjectiveScoreCategory> subjectiveCategories) {